pr fixes..
