+++ Runtime/Core/MonoSingletonGeneric.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2f0dc9e9b967de448af89c0e46c3bb92
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
