+++ Runtime/Core/MonoSingletonGeneric/ScriptExplainer.txt
