+++ Runtime/Core/MonoSingletonGeneric/ScriptExplainer.txt.meta
