+++ Runtime/Utils/TestUtils.cs
@@ -1,0 +1,12 @@
+using GameplayMechanicsUMFOSS.Core;
+using UnityEngine;
+
+public class TestUtils : MonoBehaviour
+{
+    private void Start()
+    {
+        Utils.LoadAnyLvl(1);
+    }
+
+
+}
