+++ Runtime/Utils/UtilsScriptExplainer.txt.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: fa739aea925364a4dbe874f31afe214f
-folderAsset: yes
-DefaultImporter:
+guid: 4584019396a038e4abb4a31bb25cee1c
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
